Refactored code to send Exchanges instead of just buffers.
Similar to Apache Camel, an Exchange contains a request slice (that
should be forwarded to destination if there are no errors) and a
response slice, which contains a response to be send directly to sender.

Switched to using ranges for some operations such as encryption - to
allow growing the original slices, and solve &mut borrows.
This also reduced the number of lifetimes.

Addressed some TODO FUTUREHEAP and TODO UNSHIFT comments - AWS LS RS
does copy_withing and returns statically sized arrays, so in practice
this is less impactful than expected.